requirements.txt
streamlit
numpy
opencv-contrib-python
opencv-python
Pillow
